+++ packages/ui/core-components/src/lib/unsorted/viz/index.js
@@ -11,6 +11,7 @@ export { default as CalendarHeatmap } from './heatmap/CalendarHeatmap.svelte';
 export { default as Chart } from './core/Chart.svelte';
 export { default as Column } from './table/Column.svelte';
 export { default as DataTable } from './table/DataTable.svelte';
+export { default as Delta } from './core/Delta.svelte';
 export { default as EChartsCopyTarget } from './core/EChartsCopyTarget.svelte';
 export { default as EChartsMap } from './map/EChartsMap.svelte';
 export { default as ECharts } from './core/ECharts.svelte';