+++ sites/test-env/CHANGELOG.md
@@ -1,5 +1,29 @@
 # evidence-test-environment
 
+## 3.0.24
+
+### Patch Changes
+
+- Updated dependencies [8b636a62]
+- Updated dependencies [24cec2ec]
+  - @evidence-dev/core-components@3.7.1
+  - @evidence-dev/evidence@31.0.1
+
+## 3.0.23
+
+### Patch Changes
+
+- Updated dependencies [58880cc9]
+- Updated dependencies [4f918c71]
+- Updated dependencies [d3847df4]
+- Updated dependencies [aa5708f0]
+- Updated dependencies [d97d75e6]
+- Updated dependencies [8efccce0]
+- Updated dependencies [dbf8db42]
+  - @evidence-dev/core-components@3.7.0
+  - @evidence-dev/component-utilities@2.5.1
+  - @evidence-dev/evidence@31.0.0
+
 ## 3.0.22
 
 ### Patch Changes